Remove pipe
